requirements.txt
JPype1-py3==0.5.5.4
pip==19.0.3
setuptools==40.8.0